se completan modulos empresa y sucursales

+++ src/main/java/com/umg/charly/nomina/Repository/CompanyRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CompanyRepository extends JpaRepository<Company,Long> {
 
-    public Company findByIdCompany(int idCompany);
+    public Company findByIdCompany(long idCompany);
 }
